Offline Imagery: Update file download worker
Updated the file download worker to operate directly on tile objects
rather than primitives (filename, url, etc.). This simplifies the worker
for more complex cases (such as resuming an interrupted download) in
addition to ensuring we complete the job by writing to the db even if
the application stops during the worker's processing.