perf: use specific notify events in MapCardAdapter
The docs for NotifyDataSetChanged state:

> If you are writing an adapter it will always be more efficient to use the more specific change events if you can. Rely on notifyDataSetChanged() as a last resort.

Instead of using NotifyDataSetChanged in `focusItemAtIndex`, I've switched this call to two calls to notifyItemChanged on precisely the cards that actually are affected by the call. Anecdotally, this seems to make the performance of my emulator when swiping cards much snappier, so it seems to bear out the description in NotifyDataSetChanged.

I've retained the data set call in updateData, as there we actually are resetting the entire dataset for the adapter.